Fix typo in query (#44)
I think we need to rework this one completely. The title is "Ocular query language", but it isn't actually about the query language, it's all about the Ocular console. `loadSp` is now deprecated and `loadCpg` shouldn't be used in this way anymore. I'm considering rewriting this entirely, what do you think?