remove reference to specific orb version. (#166)
Not a required change but is recommended unless the documentation will be updated to stay in sync with the current orb version. We like to leave this area open in examples so users can ensure they are getting the newest version available from the Orb registry rather than the version that was available at the time the example was created.